Add cut-down version of withinBounds()
The new logic of using 13- and 3- neighbourhoods means that checking for
out of bounds forwards (i.e. >= w, h, or d) is redundant. Additionally,
only one out of bounds case occurs where z < 0, which is scanning the
0th slice with a 13- or 3- neighbourhood, which is easily avoided in the
chunk stitching stage - i.e. don't stitch using the 0th slice of the 0th
chunk - there is nothing to which to stitch it.